refactor(testrunner): Group all IT test runners
* Group all integration test test runners for maintanability
