--- test/integration/isolated-runner/CoverageReportingTestRunner.ts
@@ -1,14 +1,0 @@
-import { EventEmitter } from 'events'; 
-import { RunResult, RunStatus, RunOptions, TestRunner, TestRunnerFactory } from 'stryker-api/test_runner';
-
-class CoverageReportingTestRunner extends EventEmitter implements TestRunner {
-
-  runResult: RunResult = { status: RunStatus.Complete, tests: [], coverage: <any>'realCoverage' };
-
-  run(options: RunOptions) {
-    (Function('return this'))().__coverage__ = 'overriden';
-    return Promise.resolve(this.runResult);
-  }
-}
-
-TestRunnerFactory.instance().register('coverage-reporting', CoverageReportingTestRunner);
