+++ test/integration/isolated-runner/IsolatedTestRunnerAdapterSpec.ts
@@ -12,15 +12,7 @@ describe('TestRunnerChildProcessAdapter', function () {
   let sut: TestRunnerChildProcessAdapter;
   let options: IsolatedRunnerOptions = {
     strykerOptions: {
-      plugins: [
-        '../../test/integration/isolated-runner/DirectResolvedTestRunner',
-        '../../test/integration/isolated-runner/NeverResolvedTestRunner',
-        '../../test/integration/isolated-runner/SlowInitAndDisposeTestRunner',
-        '../../test/integration/isolated-runner/CoverageReportingTestRunner',
-        '../../test/integration/isolated-runner/ErroredTestRunner',
-        '../../test/integration/isolated-runner/VerifyWorkingFolderTestRunner',
-        '../../test/integration/isolated-runner/CrashingTestRunner',
-        '../../test/integration/isolated-runner/DiscoverRegexTestRunner'],
+      plugins: ['../../test/integration/isolated-runner/AdditionalTestRunners'],
       testRunner: 'karma',
       testFramework: 'jasmine',
       port: null,
@@ -31,7 +23,7 @@ describe('TestRunnerChildProcessAdapter', function () {
     sandboxWorkingFolder: path.resolve('./test/integration/isolated-runner')
   };
 
-  describe.only('when test runner is crashing', () => {
+  xdescribe('when test runner is crashing', () => {
     before(() => sut = new TestRunnerChildProcessAdapter('crashing', options));
 
     it('should not give a channel error', () => {
@@ -51,7 +43,7 @@ describe('TestRunnerChildProcessAdapter', function () {
   describe('when test runner behind reports coverage', () => {
     before(() => sut = new TestRunnerChildProcessAdapter('coverage-reporting', options));
 
-    it('should not be overriden by the worker',
+    it('should not be overridden by the worker',
       () => expect(sut.run({ timeout: 3000 })).to.eventually.have.property('coverage', 'realCoverage'));
 
     after(() => sut.dispose());
